+++ packages/games/wordle/src/components/header/Header.tsx
@@ -1,24 +1,22 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import styles from "./Header.module.css";
 import { faCircleQuestion, faRotateRight } from "@fortawesome/free-solid-svg-icons";
-import { useGameContext } from "../../hooks/gameContext";
 
 interface HeaderProps {
 	toggleHelp: Function;
+	restartGame: Function;
 }
 
-export function Header({ toggleHelp }: HeaderProps) {
-	const game = useGameContext();
-
-	return <header>
-		<span className={styles["Menu-left"]}>
-			<button title="Restart" onClick={() => { game?.restartGame(); }}>
+export function Header({ toggleHelp, restartGame }: HeaderProps) {
+	return <header className={styles.Header}>
+		<span className={`${styles.Menu} ${styles["Menu-left"]}`}>
+			<button title="Restart" className={styles["Menu-button"]} onClick={() => { restartGame(); }}>
 				<FontAwesomeIcon icon={faRotateRight}/>
 			</button>
 		</span>
-		<h1>Wordle</h1>
-		<span className={styles["Menu-right"]}>
-			<button title="Show/Hide Help" onClick={() => { toggleHelp(); }}>
+		<h1 className={styles.Title}>Wordle</h1>
+		<span className={`${styles.Menu} ${styles["Menu-right"]}`}>
+			<button title="Show/Hide Help" className={styles["Menu-button"]} onClick={() => { toggleHelp(); }}>
 				<FontAwesomeIcon icon={faCircleQuestion}/>
 			</button>
 		</span>