+++ packages/games/wordle/src/components/play-area/grid/row/Row.tsx
@@ -1,22 +1,23 @@
-import { useClassNames } from "prozilla-os";
+import { useClassNames } from "@prozilla-os/core";
+import { Cell } from "./cell/Cell";
 import styles from "./Row.module.css";
-import { useGameContext } from "../../../../hooks/gameContext";
+import { CellType } from "../../../../types/grid";
 
 interface RowProps {
-	index: number;
+	row: CellType[];
+	active: boolean;
+	activeCellIndex: number;
 }
 
-export function Row({ index }: RowProps) {
-	const game = useGameContext();
-
+export function Row({ row, active, activeCellIndex }: RowProps) {
 	const classNames = [styles.Row];
 
-	if (game.activeRowIndex == index)
+	if (active)
 		classNames.push(styles.Active);
 
 	return <ul className={useClassNames(classNames)}>
-		{game && game.getRow(index).map(({ value, correctness }, letterIndex) =>
-			<li key={letterIndex} className={styles.Cell} data-correctness={correctness}>{value}</li>
+		{row.map((cell, cellIndex) =>
+			<Cell key={cellIndex} active={active && cellIndex == activeCellIndex} cell={cell}/>
 		)}
 	</ul>;
 }